Changed to have app in an application.cpp file, added first test case
