+++ tests/Group-10/Makefile
@@ -1,7 +1,7 @@
-TEST_NAMES = TestComboManager TestPermutationManager TestStaticVector TestMemoFunction TestBruteForceOptimizer TestMemoryFactory
-
+#TEST_NAMES = TestComboManager TestPermutationManager TestStaticVector TestMemoFunction TestBruteForceOptimizer TestMemoryFactory
+TEST_NAMES = TestFinalApp
 # -O3	  -Wl,--stack,8388608 -ftrack-macro-expansion=0
-FLAGS = -std=c++20 -g -pthread -Wall -Wextra -I../../ -DCATCH_CONFIG_MAIN
+FLAGS = -std=c++23 -g -pthread -Wall -Wextra -I../../ -DCATCH_CONFIG_MAIN
 
 default: test
