+++ tests/Group-10/script1.txt
@@ -1,0 +1,1 @@
+brute-force ../../Group-10/Application/example1.txt -capacity=10.0
